fix(ci): split orchestration router modules
